crypto_primes::generic

Function sieve_and_find

pub fn sieve_and_find<R, S>(
    rng: &mut R,
    sieve_factory: S,
    predicate: impl Fn(&mut R, &S::Item) -> bool,
) -> Result<Option<S::Item>, Error>
where
    S: SieveFactory,
    R: CryptoRng + ?Sized,

Sieves through the results of sieve_factory and returns the first item for which predicate is true.

If sieve_factory signals that no more results can be created, returns None.