crypto_primes

Module hazmat

Components to build your own primality test. Handle with care.

Modules

float 🔒

Const-context floating point functions that are currently not present in core.

gcd 🔒

jacobi 🔒

Jacobi symbol calculation.

libm 🔒

Functions extracted from libm to avoid a dependency on it (which sometimes breaks builds due to linking to C libraries).

lucas 🔒

Lucas primality test.

miller_rabin 🔒

Miller-Rabin primality test.

precomputed 🔒

primecount 🔒

sieve 🔒

An iterator for weeding out multiples of small primes, before proceeding with slower tests.

utils 🔒

Structs

AStarBase

“Method A*” for selecting the base given in Baillie & Wagstaff.

BruteForceBase

“Method C” for selecting the base given by Baillie.

MillerRabin

Precomputed data used to perform Miller-Rabin primality test.

SelfridgeBase

“Method A” for selecting the base given in Baillie & Wagstaff, attributed to Selfridge.

SmallFactorsSieve

An iterator returning numbers with up to and including given bit length, starting from a given number, that are not multiples of the first 2048 small primes.

SmallFactorsSieveFactory

A sieve returning numbers that are not multiples of a set of small factors.

Enums

LucasCheck

The checks to perform in the Lucas test.

Primality

Possible results of various primality tests.

SetBits

Decide how prime candidates are manipulated by setting certain bits before primality testing, influencing the range of the prime.

Traits

LucasBase

A method for selecting the base (P, Q) for the Lucas primality test.

SieveFactory

A type producing sieves for random prime generation.

Functions

estimate_primecount

Estimate the number of primes smaller than x using the asymptotic expansion of Li(x) with 4 terms, i.e.:

lucas_test

Performs the primality test based on Lucas sequence.

minimum_mr_iterations

Returns the minimum number of Miller-Rabin tests with random bases required for the given bound on the probability of hitting a pseudoprime.

random_odd_integer

Returns a random odd integer up to the given bit length.