crypto_primes::hazmat

Module float

Const-context floating point functions that are currently not present in core.

Functions

floor_sqrt 🔒

Calculates floor(sqrt(x)).

ln 🔒

pow 🔒

Calculates base^exp.

two_powf_normalized_lower_bound 🔒

Calculates a lower bound approximation of 2^exp where 0 <= exp <= 1.

two_powf_upper_bound 🔒

Calculates an approximation of 2^exp where exp < 0. The approximation is guaranteed to always be greater than 2^exp.

two_powi 🔒

Calculates 2^exp.