crypto_primes::hazmat

Module jacobi

Jacobi symbol calculation.

Functions

apply_reduce_numerator 🔒

Transforms (a/p) -> (r/p) for odd p, where the resulting r is odd, and a = r * 2^s. Takes a Jacobi symbol value, and returns r and the new Jacobi symbol, negated if the transformation changes parity.

apply_swap 🔒

Transforms (a/p) -> (p/a) for odd and coprime a and p. Takes a Jacobi symbol value, and returns the swapped pair and the new Jacobi symbol, negated if the transformation changes parity.

jacobi_symbol_vartime 🔒

Returns the Jacobi symbol (a/p) given an odd p.

reduce_numerator_long 🔒

reduce_numerator_short 🔒

swap_long 🔒

swap_short 🔒