crypto_primes::hazmat

Module libm

Functions extracted from libm to avoid a dependency on it (which sometimes breaks builds due to linking to C libraries).

Some reformatting and clippy exceptions were added.

Constants

F64_BITS 🔒

The bitwidth of the float type.

F64_EXP_BIAS 🔒

The exponent bias value.

F64_EXP_BITS 🔒

The bitwidth of the exponent.

F64_EXP_SAT 🔒

The saturated (maximum bitpattern) value of the exponent, i.e. the infinite representation.

F64_SIGN_MASK 🔒

A mask for the sign bit

F64_SIG_BITS 🔒

The bitwidth of the significand (does not include the implicit bit).

F64_SIG_MASK 🔒

A mask for the significand

IVLN2HI 🔒

IVLN2LO 🔒

LG1 🔒

LG2 🔒

LG3 🔒

LG4 🔒

LG5 🔒

LG6 🔒

LG7 🔒

Functions

f64_ex 🔒

Returns the exponent, not adjusting for bias, not accounting for subnormals or zero.

f64_exp_unbiased 🔒

Extract the exponent and adjust it for bias, not accounting for subnormals or zero.

f64_from_parts 🔒

Constructs an f64 from its parts. Inputs are treated as bits and shifted into position.

floor 🔒

Generic floor algorithm.

log2 🔒

The base 2 logarithm of x (f64).

round 🔒

trunc 🔒

Type Aliases

F64Int 🔒

A uint of the same width as the float