crypto_primes::hazmat

Module lucas

Lucas primality test.

Structs

AStarBase

“Method A*” for selecting the base given in Baillie & Wagstaff.

BruteForceBase

“Method C” for selecting the base given by Baillie.

SelfridgeBase

“Method A” for selecting the base given in Baillie & Wagstaff, attributed to Selfridge.

Enums

LucasCheck

The checks to perform in the Lucas test.

Constants

ATTEMPTS_BEFORE_SQRT 🔒

The number of attempts to find D such that (D/n) == -1 before checking that n is a square (in which case such D does not exist).

MAX_ATTEMPTS 🔒

The maximum number of attempts to find D such that (D/n) == -1.

Traits

LucasBase

A method for selecting the base (P, Q) for the Lucas primality test.

Functions

decompose 🔒

For the given odd n, finds s and odd d such that n + 1 == 2^s * d.

lucas_test

Performs the primality test based on Lucas sequence.