crypto_primes::hazmat::miller_rabin

Function pseudoprime_probability

const fn pseudoprime_probability(k: u32, t: u32, cap_m: u32) -> f64

Returns the probability p_{k,t} of an odd k-bit integer passing t rounds of MR testing with random bases is actually composite.

Taken from FIPS-186.51, Section C.1, Eq. (2); FIPS in turn quotes Damgård et al2. There it can be found as Eq. (4.1), with the factors bounded by Proposition 1 and Proposition 2 in Section 4.

Warning:1 Care must be taken when using the phrase “error probability” in connection with the recommended number of rounds of MR testing. The probability that a composite number of bit length k survives t rounds of Miller-Rabin testing is not the same as p_{k,t}, which is the probability that a number surviving t rounds of Miller-Rabin testing is composite. Ordinarily, the latter probability is the one that should be of most interest to a party responsible for generating primes, while the former may be more important to a party responsible for validating the primality of a number generated by someone else. However, for sufficiently large k (e.g., k ≥ 51), it can be shown that p_{k,t} ≤ 4^{–t} under the same assumptions concerning the selection of candidates as those made to obtain formula for p_{k,t} (see Damgård et al2). In such cases, t = ceil(–log2(p_target)/2) rounds of Miller-Rabin testing can be used to both generate and validate primes with p_target serving as an upper bound on both the probability that the generation process yields a composite number and the probability that a composite number would survive an attempt to validate its primality.

FIPS-186.5 standard, https://nvlpubs.nist.gov/nistpubs/FIPS/NIST.FIPS.186-5.pdf ↩ 1 2

Damgård I, Landrock P, Pomerance C (1993) Average Case Error Estimates for the Strong Probable Prime Test. Mathematics of Computation 61(203):177-194 (1993), https://www.ams.org/journals/mcom/1993-61-203/S0025-5718-1993-1189518-9/S0025-5718-1993-1189518-9.pdf ↩ 1 2