crypto_primes::hazmat::sieve

Function trial_primes_num

fn trial_primes_num<T>(start: &T, max_bit_length: NonZeroU32) -> usize
where
    T: Unsigned,

Returns the number of trial prime factors from SMALL_PRIMES to use for trial division of the candidates in range [start, 2^max_bit_length).

None of the factors chosen this way will be greater or equal to start (this is important for the use in the sieve, because when we only have the residue, we cannot distinguish between a prime itself and a multiple of that prime).