crypto_primes::presets

Function is_prime

pub fn is_prime<T>(flavor: Flavor, candidate: &T) -> bool
where
    T: UnsignedWithMontyForm + RandomMod,

Checks if the given number is prime.

Performed tests:

Miller-Rabin test with base 2;

LucasCheck::Bpsw21 test with AStarBase.

See MillerRabin and lucas_test for more details about the tests.

This is the recommended approach by Baillie et al1, improving on the BPSW’80 test2. At the moment of the writing there are no known composites that are reported to be primes by either of the approaches1; it is conjectured that they may exist, but their size is larger than the numbers that are used in practice.

R. Baillie, A. Fiori, S. S. Wagstaff, “Strengthening the Baillie-PSW primality test”, Math. Comp. 90 1931-1955 (2021), DOI: 10.1090/mcom/3616 ↩ 1 2

R. Baillie, S. S. Wagstaff, “Lucas pseudoprimes”, Math. Comp. 35 1391-1417 (1980), DOI: 10.2307/2006406, http://mpqs.free.fr/LucasPseudoprimes.pdf ↩