crypto_primes::hazmat

Function minimum_mr_iterations

pub const fn minimum_mr_iterations(
    bit_length: u32,
    log2_target: u32,
) -> Option<usize>

For a candidate of size bit_length, returns the minimum number of Miller-Rabin tests with random bases required for the probability of hitting a pseudoprime (that is, a composite for which all M-R tests pass) to be smaller than 2^{-log2_target}.

Returns None if the number of iterations could not be found for the given bounds.

This function implements the formula prescribed by the FIPS.186-5 standard1.

FIPS-186.5 standard, https://nvlpubs.nist.gov/nistpubs/FIPS/NIST.FIPS.186-5.pdf ↩